requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py[voice]
asyncpg
python-dateutil
psutil
